ZCS-806 Change scope of sieve ldap attributes
ZCS-806 Change as per review comment
