+++ store/src/java-test/com/zimbra/cs/filter/NotifyMailtoTest.java
@@ -33,8 +33,8 @@ import org.junit.Test;
 import com.google.common.collect.Maps;
 import com.zimbra.common.account.Key;
 import com.zimbra.common.util.ArrayUtil;
-import com.zimbra.common.util.ByteUtil;
 import com.zimbra.cs.account.Account;
+import com.zimbra.cs.account.MockProvisioning;
 import com.zimbra.cs.account.Provisioning;
 import com.zimbra.cs.account.Server;
 import com.zimbra.cs.lmtpserver.LmtpAddress;
@@ -58,12 +58,13 @@ public class NotifyMailtoTest {
     public static void init() throws Exception {
         MailboxTestUtil.initServer();
         MailboxTestUtil.clearData();
-        Provisioning prov = Provisioning.getInstance();
+        MockProvisioning prov = new MockProvisioning();
+        Provisioning.setInstance(prov);
 
         Map<String, Object> attrs = Maps.newHashMap();
         attrs = Maps.newHashMap();
         attrs.put(Provisioning.A_zimbraMailSieveNotifyActionRFCCompliant, "TRUE");
-        prov.getConfig().modify(attrs);
+        prov.getLocalServer().modify(attrs);
 
         prov.createDomain("zimbra.com", attrs);
