+++ store/src/java/com/zimbra/cs/account/ZAttrConfig.java
@@ -25345,314 +25345,6 @@ public abstract class ZAttrConfig extends Entry {
     }
 
     /**
-     * outgoing sieve script defined by admin (not able to edit and view from
-     * the end user) applied after the end user filter rule
-     *
-     * @return zimbraMailAdminOutgoingSieveScriptAfter, or null if unset
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2093)
-    public String getMailAdminOutgoingSieveScriptAfter() {
-        return getAttr(Provisioning.A_zimbraMailAdminOutgoingSieveScriptAfter, null, true);
-    }
-
-    /**
-     * outgoing sieve script defined by admin (not able to edit and view from
-     * the end user) applied after the end user filter rule
-     *
-     * @param zimbraMailAdminOutgoingSieveScriptAfter new value
-     * @throws com.zimbra.common.service.ServiceException if error during update
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2093)
-    public void setMailAdminOutgoingSieveScriptAfter(String zimbraMailAdminOutgoingSieveScriptAfter) throws com.zimbra.common.service.ServiceException {
-        HashMap<String,Object> attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptAfter, zimbraMailAdminOutgoingSieveScriptAfter);
-        getProvisioning().modifyAttrs(this, attrs);
-    }
-
-    /**
-     * outgoing sieve script defined by admin (not able to edit and view from
-     * the end user) applied after the end user filter rule
-     *
-     * @param zimbraMailAdminOutgoingSieveScriptAfter new value
-     * @param attrs existing map to populate, or null to create a new map
-     * @return populated map to pass into Provisioning.modifyAttrs
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2093)
-    public Map<String,Object> setMailAdminOutgoingSieveScriptAfter(String zimbraMailAdminOutgoingSieveScriptAfter, Map<String,Object> attrs) {
-        if (attrs == null) attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptAfter, zimbraMailAdminOutgoingSieveScriptAfter);
-        return attrs;
-    }
-
-    /**
-     * outgoing sieve script defined by admin (not able to edit and view from
-     * the end user) applied after the end user filter rule
-     *
-     * @throws com.zimbra.common.service.ServiceException if error during update
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2093)
-    public void unsetMailAdminOutgoingSieveScriptAfter() throws com.zimbra.common.service.ServiceException {
-        HashMap<String,Object> attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptAfter, "");
-        getProvisioning().modifyAttrs(this, attrs);
-    }
-
-    /**
-     * outgoing sieve script defined by admin (not able to edit and view from
-     * the end user) applied after the end user filter rule
-     *
-     * @param attrs existing map to populate, or null to create a new map
-     * @return populated map to pass into Provisioning.modifyAttrs
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2093)
-    public Map<String,Object> unsetMailAdminOutgoingSieveScriptAfter(Map<String,Object> attrs) {
-        if (attrs == null) attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptAfter, "");
-        return attrs;
-    }
-
-    /**
-     * outgoing sieve script defined by admin (not able to edit and view from
-     * the end user) applied before the end user filter rule
-     *
-     * @return zimbraMailAdminOutgoingSieveScriptBefore, or null if unset
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2092)
-    public String getMailAdminOutgoingSieveScriptBefore() {
-        return getAttr(Provisioning.A_zimbraMailAdminOutgoingSieveScriptBefore, null, true);
-    }
-
-    /**
-     * outgoing sieve script defined by admin (not able to edit and view from
-     * the end user) applied before the end user filter rule
-     *
-     * @param zimbraMailAdminOutgoingSieveScriptBefore new value
-     * @throws com.zimbra.common.service.ServiceException if error during update
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2092)
-    public void setMailAdminOutgoingSieveScriptBefore(String zimbraMailAdminOutgoingSieveScriptBefore) throws com.zimbra.common.service.ServiceException {
-        HashMap<String,Object> attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptBefore, zimbraMailAdminOutgoingSieveScriptBefore);
-        getProvisioning().modifyAttrs(this, attrs);
-    }
-
-    /**
-     * outgoing sieve script defined by admin (not able to edit and view from
-     * the end user) applied before the end user filter rule
-     *
-     * @param zimbraMailAdminOutgoingSieveScriptBefore new value
-     * @param attrs existing map to populate, or null to create a new map
-     * @return populated map to pass into Provisioning.modifyAttrs
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2092)
-    public Map<String,Object> setMailAdminOutgoingSieveScriptBefore(String zimbraMailAdminOutgoingSieveScriptBefore, Map<String,Object> attrs) {
-        if (attrs == null) attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptBefore, zimbraMailAdminOutgoingSieveScriptBefore);
-        return attrs;
-    }
-
-    /**
-     * outgoing sieve script defined by admin (not able to edit and view from
-     * the end user) applied before the end user filter rule
-     *
-     * @throws com.zimbra.common.service.ServiceException if error during update
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2092)
-    public void unsetMailAdminOutgoingSieveScriptBefore() throws com.zimbra.common.service.ServiceException {
-        HashMap<String,Object> attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptBefore, "");
-        getProvisioning().modifyAttrs(this, attrs);
-    }
-
-    /**
-     * outgoing sieve script defined by admin (not able to edit and view from
-     * the end user) applied before the end user filter rule
-     *
-     * @param attrs existing map to populate, or null to create a new map
-     * @return populated map to pass into Provisioning.modifyAttrs
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2092)
-    public Map<String,Object> unsetMailAdminOutgoingSieveScriptBefore(Map<String,Object> attrs) {
-        if (attrs == null) attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptBefore, "");
-        return attrs;
-    }
-
-    /**
-     * sieve script defined by admin (not able to edit and view from the end
-     * user) applied after the end user filter rule
-     *
-     * @return zimbraMailAdminSieveScriptAfter, or null if unset
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2091)
-    public String getMailAdminSieveScriptAfter() {
-        return getAttr(Provisioning.A_zimbraMailAdminSieveScriptAfter, null, true);
-    }
-
-    /**
-     * sieve script defined by admin (not able to edit and view from the end
-     * user) applied after the end user filter rule
-     *
-     * @param zimbraMailAdminSieveScriptAfter new value
-     * @throws com.zimbra.common.service.ServiceException if error during update
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2091)
-    public void setMailAdminSieveScriptAfter(String zimbraMailAdminSieveScriptAfter) throws com.zimbra.common.service.ServiceException {
-        HashMap<String,Object> attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptAfter, zimbraMailAdminSieveScriptAfter);
-        getProvisioning().modifyAttrs(this, attrs);
-    }
-
-    /**
-     * sieve script defined by admin (not able to edit and view from the end
-     * user) applied after the end user filter rule
-     *
-     * @param zimbraMailAdminSieveScriptAfter new value
-     * @param attrs existing map to populate, or null to create a new map
-     * @return populated map to pass into Provisioning.modifyAttrs
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2091)
-    public Map<String,Object> setMailAdminSieveScriptAfter(String zimbraMailAdminSieveScriptAfter, Map<String,Object> attrs) {
-        if (attrs == null) attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptAfter, zimbraMailAdminSieveScriptAfter);
-        return attrs;
-    }
-
-    /**
-     * sieve script defined by admin (not able to edit and view from the end
-     * user) applied after the end user filter rule
-     *
-     * @throws com.zimbra.common.service.ServiceException if error during update
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2091)
-    public void unsetMailAdminSieveScriptAfter() throws com.zimbra.common.service.ServiceException {
-        HashMap<String,Object> attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptAfter, "");
-        getProvisioning().modifyAttrs(this, attrs);
-    }
-
-    /**
-     * sieve script defined by admin (not able to edit and view from the end
-     * user) applied after the end user filter rule
-     *
-     * @param attrs existing map to populate, or null to create a new map
-     * @return populated map to pass into Provisioning.modifyAttrs
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2091)
-    public Map<String,Object> unsetMailAdminSieveScriptAfter(Map<String,Object> attrs) {
-        if (attrs == null) attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptAfter, "");
-        return attrs;
-    }
-
-    /**
-     * sieve script defined by admin (not able to edit and view from the end
-     * user) applied before the end user filter rule
-     *
-     * @return zimbraMailAdminSieveScriptBefore, or null if unset
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2090)
-    public String getMailAdminSieveScriptBefore() {
-        return getAttr(Provisioning.A_zimbraMailAdminSieveScriptBefore, null, true);
-    }
-
-    /**
-     * sieve script defined by admin (not able to edit and view from the end
-     * user) applied before the end user filter rule
-     *
-     * @param zimbraMailAdminSieveScriptBefore new value
-     * @throws com.zimbra.common.service.ServiceException if error during update
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2090)
-    public void setMailAdminSieveScriptBefore(String zimbraMailAdminSieveScriptBefore) throws com.zimbra.common.service.ServiceException {
-        HashMap<String,Object> attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptBefore, zimbraMailAdminSieveScriptBefore);
-        getProvisioning().modifyAttrs(this, attrs);
-    }
-
-    /**
-     * sieve script defined by admin (not able to edit and view from the end
-     * user) applied before the end user filter rule
-     *
-     * @param zimbraMailAdminSieveScriptBefore new value
-     * @param attrs existing map to populate, or null to create a new map
-     * @return populated map to pass into Provisioning.modifyAttrs
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2090)
-    public Map<String,Object> setMailAdminSieveScriptBefore(String zimbraMailAdminSieveScriptBefore, Map<String,Object> attrs) {
-        if (attrs == null) attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptBefore, zimbraMailAdminSieveScriptBefore);
-        return attrs;
-    }
-
-    /**
-     * sieve script defined by admin (not able to edit and view from the end
-     * user) applied before the end user filter rule
-     *
-     * @throws com.zimbra.common.service.ServiceException if error during update
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2090)
-    public void unsetMailAdminSieveScriptBefore() throws com.zimbra.common.service.ServiceException {
-        HashMap<String,Object> attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptBefore, "");
-        getProvisioning().modifyAttrs(this, attrs);
-    }
-
-    /**
-     * sieve script defined by admin (not able to edit and view from the end
-     * user) applied before the end user filter rule
-     *
-     * @param attrs existing map to populate, or null to create a new map
-     * @return populated map to pass into Provisioning.modifyAttrs
-     *
-     * @since ZCS 8.7.8
-     */
-    @ZAttr(id=2090)
-    public Map<String,Object> unsetMailAdminSieveScriptBefore(Map<String,Object> attrs) {
-        if (attrs == null) attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptBefore, "");
-        return attrs;
-    }
-
-    /**
      * Whether to allow password sent to non-secured port when zimbraMailMode
      * is mixed. If it set to TRUE the server will allow login with clear
      * text AuthRequests and change password with clear text
